backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asgiref>=3.8.1",
    "django-cors-headers>=4.7.0",
    "django>=5.1.6",
    "djangorestframework>=3.15.2",
    "djangorestframework-simplejwt>=5.4.0",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.0.1",
    "pytz>=2025.1",
    "sqlparse>=0.5.3",
    "django-money>=3.5.3",
    "loguru>=0.7.3",
]
